Added current landing page HTML and CSS
The join the alpha link is broken currently. I want to work on getting it to just append a list of emails in a spreadsheet somehow. I've been looking into the Google Sheets API and it doesn't seem THAT difficult but I have read quite a bit of words and I would argue at least 9X that amount in letters alone. So I'm tired and will get to it another day.